Restart
Decided to actually finish this off now. So what i've done is i've taken the original script given at mediapipe and changed it up so that every 5 seconds it grabs all the cordinates of the hand and puts them into a list with the sign i wanted to show at that time.

The sign is given via a generator which loops through the numbers 0-2 and using that gives the index of the sign to be used.

I've chosen 3 signs to do, all single hand actions and are static so no need to implement video for decision making. I will say i need to change one of the words, but that will be done tomorrow when i start to implement torch for convolutional neural networks. I'd use keras but they no longer support GPU on windows for some fucking reason.